+++ src/backend/joanie/tests/test_api_course.py
@@ -92,7 +92,7 @@ class CourseApiTest(BaseAPITestCase):
                 if should_enroll:
                     course_run = random.choice(target_course_runs)
                     factories.EnrollmentFactory(
-                        user=user, course_run=course_run, order=order, is_active=True
+                        user=user, course_run=course_run, is_active=True
                     )
 
         with self.assertNumQueries(11):
@@ -257,11 +257,11 @@ class CourseApiTest(BaseAPITestCase):
 
         # - Enrollment to course run 11
         factories.EnrollmentFactory(
-            user=user, course_run=target_course_run11, order=order1, is_active=True
+            user=user, course_run=target_course_run11, is_active=True
         )
         # - Enrollment to course run 12
         factories.EnrollmentFactory(
-            user=user, course_run=target_course_run12, order=order1, is_active=True
+            user=user, course_run=target_course_run12, is_active=True
         )
         # - Product 2
         order2 = factories.OrderFactory(
@@ -273,11 +273,11 @@ class CourseApiTest(BaseAPITestCase):
         ProformaInvoiceFactory(order=order2, total=order2.total)
         # - Enrollment to course run 21
         factories.EnrollmentFactory(
-            user=user, course_run=target_course_run21, order=order2, is_active=True
+            user=user, course_run=target_course_run21, is_active=True
         )
         # - Enrollment to course run 22
         factories.EnrollmentFactory(
-            user=user, course_run=target_course_run22, order=order2, is_active=True
+            user=user, course_run=target_course_run22, is_active=True
         )
 
         # - Create a certificate
@@ -308,10 +308,10 @@ class CourseApiTest(BaseAPITestCase):
                         )
                     )
                     factories.EnrollmentFactory(
-                        user=user, course_run=course_run, order=order, is_active=True
+                        user=user, course_run=course_run, is_active=True
                     )
 
-        with self.assertNumQueries(29):
+        with self.assertNumQueries(35):
             response = self.client.get(
                 f"/api/courses/{course.code}/",
                 HTTP_AUTHORIZATION=f"Bearer {token}",
@@ -357,7 +357,7 @@ class CourseApiTest(BaseAPITestCase):
                                 "+00:00", "Z"
                             ),
                         }
-                        for enrollment in order.enrollments.all()
+                        for enrollment in order.get_enrollments()
                     ],
                 }
                 for order in [order1, order2]
@@ -432,7 +432,7 @@ class CourseApiTest(BaseAPITestCase):
 
         # - When user is authenticated, response should be partially cached.
         # Course information should have been cached, but orders not.
-        with self.assertNumQueries(10):
+        with self.assertNumQueries(16):
             self.client.get(
                 f"/api/courses/{course.code}/",
                 HTTP_AUTHORIZATION=f"Bearer {token}",
@@ -478,7 +478,7 @@ class CourseApiTest(BaseAPITestCase):
         self.assertEqual(order_canceled.state, enums.ORDER_STATE_CANCELED)
 
         # - Retrieve course information
-        with self.assertNumQueries(13):
+        with self.assertNumQueries(16):
             response = self.client.get(
                 f"/api/courses/{course.code}/",
                 HTTP_AUTHORIZATION=f"Bearer {token}",
@@ -534,7 +534,6 @@ class CourseApiTest(BaseAPITestCase):
         factories.EnrollmentFactory(
             course_run=cr1,
             user=user,
-            order=order,
             is_active=True,
             state=enums.ENROLLMENT_STATE_SET,
         )